backend/requirements.txt
blinker==1.6.2
click==8.1.3
colormath @ git+https://github.com/codedbychavez/python-colormath
Flask==2.3.2
Flask-Cors==3.0.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
six==1.16.0
smartcrop==0.3.3
Werkzeug==2.3.3
zipp==3.15.0
